Files/src/trading_engine.py
"""
Advanced Quantitative Trading Platform
=====================================

Institutional-grade algorithmic trading system with deep learning models,
real-time market analysis, and sophisticated risk management capabilities.

Author: Emilio Cardenas
License: MIT
"""

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Advanced ML and Deep Learning
import tensorflow as tf
from tensorflow.keras.models import Sequential, Model
from tensorflow.keras.layers import LSTM, Dense, Dropout, BatchNormalization, Attention, MultiHeadAttention
from tensorflow.keras.layers import Input, LayerNormalization, Add, GlobalAveragePooling1D
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.callbacks import EarlyStopping, ReduceLROnPlateau, ModelCheckpoint

# Traditional ML
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
import xgboost as xgb
import lightgbm as lgb

# Financial Data and Analysis
import yfinance as yf
import pandas_ta as ta
from alpha_vantage.timeseries import TimeSeries
from alpha_vantage.techindicators import TechIndicators

# Statistical Analysis
from scipy import stats
from statsmodels.tsa.arima.model import ARIMA
from statsmodels.tsa.statespace.sarimax import SARIMAX
from arch import arch_model

# Visualization and Reporting
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots

# Risk Management
import scipy.optimize as sco
from pypfopt import EfficientFrontier, risk_models, expected_returns
from pypfopt.discrete_allocation import DiscreteAllocation, get_latest_prices

# Real-time Processing
from datetime import datetime, timedelta
import asyncio
import websocket
import json
import threading
import time

class AdvancedQuantitativeTradingEngine:
    """
    Institutional-grade quantitative trading platform with advanced AI capabilities.
    
    Features:
    - LSTM and Transformer neural networks for price prediction
    - Real-time market data processing and analysis
    - Advanced technical indicators and feature engineering
    - Portfolio optimization and risk management
    - Backtesting and performance analytics
    - Automated trading signal generation
    - Risk-adjusted return optimization
    """
    
    def __init__(self, config=None):
        """Initialize the quantitative trading engine."""
        self.config = config or self._default_config()
        self.models = {}
        self.scalers = {}
        self.market_data = {}
        self.portfolio = {}
        self.performance_metrics = {}
        self.risk_metrics = {}
        
    def _default_config(self):
        """Default configuration for the trading engine."""
        return {
            'symbols': ['AAPL', 'GOOGL', 'MSFT', 'AMZN', 'TSLA', 'NVDA', 'META', 'NFLX'],
            'lookback_period': 60,
            'prediction_horizon': 5,
            'train_test_split': 0.8,
            'lstm_units': [128, 64, 32],
            'transformer_heads': 8,
            'transformer_layers': 4,
            'learning_rate': 0.001,
            'epochs': 100,
            'batch_size': 32,
            'risk_free_rate': 0.02,
            'max_position_size': 0.1,
            'stop_loss': 0.05,
            'take_profit': 0.15,
            'rebalance_frequency': 'weekly'
        }
    
    def fetch_market_data(self, symbols=None, period='2y'):
        """Fetch comprehensive market data with advanced preprocessing."""
        symbols = symbols or self.config['symbols']
        
        print(f"Fetching market data for {len(symbols)} symbols...")
        
        for symbol in symbols:
            try:
                # Fetch OHLCV data
                ticker = yf.Ticker(symbol)
                data = ticker.history(period=period)
                
                if data.empty:
                    print(f"Warning: No data found for {symbol}")
                    continue
                
                # Advanced technical indicators
                data = self._calculate_technical_indicators(data)
                
                # Market microstructure features
                data = self._calculate_microstructure_features(data)
                
                # Volatility and risk metrics
                data = self._calculate_volatility_metrics(data)
                
                # Sentiment and momentum indicators
                data = self._calculate_sentiment_indicators(data)
                
                self.market_data[symbol] = data
                print(f"Successfully processed data for {symbol}: {len(data)} records")
                
            except Exception as e:
                print(f"Error fetching data for {symbol}: {str(e)}")
        
        return self.market_data
    
    def _calculate_technical_indicators(self, data):
        """Calculate comprehensive technical indicators."""
        # Moving averages
        data['SMA_10'] = ta.sma(data['Close'], length=10)
        data['SMA_20'] = ta.sma(data['Close'], length=20)
        data['SMA_50'] = ta.sma(data['Close'], length=50)
        data['EMA_12'] = ta.ema(data['Close'], length=12)
        data['EMA_26'] = ta.ema(data['Close'], length=26)
        
        # MACD
        macd = ta.macd(data['Close'])
        data = pd.concat([data, macd], axis=1)
        
        # RSI
        data['RSI'] = ta.rsi(data['Close'], length=14)
        
        # Bollinger Bands
        bb = ta.bbands(data['Close'], length=20)
        data = pd.concat([data, bb], axis=1)
        
        # Stochastic Oscillator
        stoch = ta.stoch(data['High'], data['Low'], data['Close'])
        data = pd.concat([data, stoch], axis=1)
        
        # Williams %R
        data['WILLR'] = ta.willr(data['High'], data['Low'], data['Close'])
        
        # Average True Range
        data['ATR'] = ta.atr(data['High'], data['Low'], data['Close'])
        
        # Commodity Channel Index
        data['CCI'] = ta.cci(data['High'], data['Low'], data['Close'])
        
        # Money Flow Index
        data['MFI'] = ta.mfi(data['High'], data['Low'], data['Close'], data['Volume'])
        
        return data
    
    def _calculate_microstructure_features(self, data):
        """Calculate market microstructure features."""
        # Price-based features
        data['Returns'] = data['Close'].pct_change()
        data['Log_Returns'] = np.log(data['Close'] / data['Close'].shift(1))
        data['High_Low_Ratio'] = data['High'] / data['Low']
        data['Open_Close_Ratio'] = data['Open'] / data['Close']
        
        # Volume-based features
        data['Volume_MA'] = data['Volume'].rolling(window=20).mean()
        data['Volume_Ratio'] = data['Volume'] / data['Volume_MA']
        data['Price_Volume'] = data['Close'] * data['Volume']
        
        # Volatility features
        data['Volatility_5'] = data['Returns'].rolling(window=5).std()
        data['Volatility_20'] = data['Returns'].rolling(window=20).std()
        
        # Gap analysis
        data['Gap'] = (data['Open'] - data['Close'].shift(1)) / data['Close'].shift(1)