requirements.txt
# ============================================================================
# Quantitative Trading Intelligence System - Requirements
# ============================================================================
# Institutional-grade algorithmic trading with deep learning models
# Author: Emilio Cardenas
# Version: 2.0.0
# ============================================================================

# Core Data Science and Analytics
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0

# Machine Learning - Core Libraries
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0

# Deep Learning and Neural Networks
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.31.0

# Advanced ML Libraries
shap>=0.42.0
optuna>=3.3.0
hyperopt>=0.2.7
stable-baselines3>=2.0.0

# Financial Data and Market APIs
yfinance>=0.2.18
alpha-vantage>=2.3.1
quandl>=3.7.0
pandas-datareader>=0.10.0
fredapi>=0.5.0
ccxt>=4.0.0
polygon-api-client>=1.12.0
twelve-data>=1.2.0

# Technical Analysis and Indicators
pandas-ta>=0.3.14b
ta-lib>=0.4.25
technical-indicators>=0.1.0
finta>=1.3.0

# Portfolio Optimization and Risk Management
pypfopt>=1.5.4
riskfolio-lib>=4.3.0
pyfolio>=0.9.2
empyrical>=0.5.5
cvxpy>=1.3.0
quantlib>=1.29

# Time Series Analysis and Forecasting
prophet>=1.1.4
pmdarima>=2.0.3
tslearn>=0.6.2
sktime>=0.21.0
arch>=5.5.0

# Web Framework and API
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=1.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0
websockets>=11.0.0

# Database and Data Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0
redis>=4.6.0
influxdb-client>=1.37.0

# Real-time Data Processing
apache-kafka-python>=2.0.0
confluent-kafka>=2.2.0
pika>=1.3.0
celery>=5.3.0

# Visualization and Dashboards
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.11.0
streamlit>=1.25.0
bokeh>=3.2.0
mplfinance>=0.12.9b7

# Power BI Integration
powerbi-client>=3.1.0
adal>=1.2.7
msal>=1.23.0

# Trading and Brokerage APIs
alpaca-trade-api>=3.1.0
ib-insync>=0.9.86
robin-stocks>=3.0.0
tda-api>=1.3.0

# Statistical Analysis and Econometrics
pymc>=5.6.0
pystan>=3.7.0
scipy-stats>=1.10.0
lifelines>=0.27.0

# Alternative Data Sources
newsapi-python>=0.2.6
twitter-api-v2>=1.4.4
beautifulsoup4>=4.12.0
scrapy>=2.10.0

# Natural Language Processing for Sentiment
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Feature Engineering and Selection
feature-engine>=1.6.0
category-encoders>=2.6.0
boruta>=0.3.0
tsfresh>=0.20.0

# Model Deployment and Monitoring
mlflow>=2.5.0
wandb>=0.15.0
neptune-client>=1.4.0
evidently>=0.4.0

# Backtesting and Strategy Testing
backtrader>=1.9.76.123
zipline>=3.0.0
vectorbt>=0.25.2
backtesting>=0.3.3

# Options and Derivatives
py_vollib>=1.0.1
mibian>=0.1.3
options-analytics>=0.1.0

# Cryptocurrency Trading
python-binance>=1.0.17
coinbase-pro>=1.0.0
kraken-api>=0.1.4

# Security and Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
dynaconf>=3.2.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing and Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0

# Code Quality and Formatting
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0

# API Documentation
swagger-ui-bundle>=0.0.9
flasgger>=0.9.7

# Caching and Performance
joblib>=1.3.0
numba>=0.57.0
cython>=3.0.0
bottleneck>=1.3.7

# Distributed Computing
dask>=2023.7.0
ray>=2.5.0

# Cloud Integration
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Profiling and Debugging
memory-profiler>=0.61.0
line-profiler>=4.0.0
py-spy>=0.3.14

# Utilities
tqdm>=4.65.0
click>=8.1.0
colorama>=0.4.6
python-dateutil>=2.8.2
pytz>=2023.3

# Data Processing and ETL
requests>=2.31.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
lxml>=4.9.0

# Market Data Visualization
cufflinks>=0.17.3
plotly-dash>=2.11.0

# Algorithmic Trading Frameworks
lean>=2.5.0
quantconnect>=2.5.0

# Development and Docker
docker>=6.1.0
docker-compose>=1.29.0

# High-Frequency Trading
pyalgotrade>=0.20
bt>=0.2.9
zipline-reloaded>=3.0.0
